server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.14.0
chromadb==0.4.18
langchain>=0.1.13
python-multipart==0.0.6
pypdf2==3.0.1
python-dotenv==1.0.0